==> Step1_Data+Extraction+and+Cleaning.ipynb <==
# First we import all the libraries required for basic data manipulation
from scipy.stats import skew
from IPython.display import display
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import pandas as pd

# We define the get_data() function which imports our datasets into the working environment
def get_data():
    """First get our csv files and import as DataFrames"""
    SP500=pd.read_csv('SP500.csv', index_col='Date', parse_dates=True)
    Nasdaq=pd.read_csv('NASDAQ.csv', index_col='Date', parse_dates=True)
    DJI=pd.read_csv('DJI.csv', index_col='Date', parse_dates=True)
    DAX=pd.read_csv('DAX.csv', index_col='Date', parse_dates=True)
    Paris=pd.read_csv('CAC40.csv', index_col='Date', parse_dates=True)
    Tokyo=pd.read_csv('N225.csv', index_col='Date', parse_dates=True)
    HongKong=pd.read_csv('HSI.csv', index_col='Date', parse_dates=True)
    Aus=pd.read_csv('ASX.csv', index_col='Date', parse_dates=True)
    """Get rid of the first year because ASX has no data"""
    SP500 = SP500[502:]
    """Fill in missing data by forward fill"""
    SP500.fillna(method='ffill',inplace=True)
    Nasdaq.fillna(method='ffill',inplace=True)
    DJI.fillna(method='ffill',inplace=True)
    DAX.fillna(method='ffill',inplace=True)
    Paris.fillna(method='ffill',inplace=True)
    Tokyo.fillna(method='ffill',inplace=True)
    HongKong.fillna(method='ffill',inplace=True)
    Aus.fillna(method='ffill',inplace=True)
    return SP500, Nasdaq, DJI, DAX, Paris, Tokyo, HongKong, Aus

# Using the above function to import datasets into the working environment
SP500, Nasdaq, DJI, DAX, Paris, Tokyo, HongKong, Aus = get_data()

# We define the left_join() function which joins the child dataset to the mother using the mother's index
def left_join(mother, child):
    """This function grabs data from all dfs on days SP500 was traded"""
    df_temp = pd.DataFrame(index = mother.index)
    df_temp1 = df_temp.join(child)
    df_temp1 = df_temp1.replace('null', np.nan)
    df_temp1.fillna(method='ffill', inplace=True)
    df_temp1.fillna(method='backfill', inplace=True)
    return df_temp1

# Using the above function to join
Nasdaq_new = left_join(SP500, Nasdaq)
DJI_new = left_join(SP500, DJI)
DAX_new = left_join(SP500, DAX)
Paris_new = left_join(SP500, Paris)
Tokyo_new = left_join(SP500, Tokyo)
HongKong_new = left_join(SP500, HongKong)
Aus_new = left_join(SP500, Aus)

# We define the reset_index() function which resets the index (previously using dates) of the dataframe
def reset_index(df):
    """Dates are no longer important"""
    df['Date'] = df.index
    df = df.reset_index(level=['Date'])
    return df

# Using the above function to reset the indexes of each dataframe
SP500 = reset_index(SP500)
Nasdaq = reset_index(Nasdaq_new)
DJI = reset_index(DJI_new)
DAX = reset_index(DAX_new)
Paris = reset_index(Paris_new)
Tokyo = reset_index(Tokyo_new)
HongKong = reset_index(HongKong_new)
Aus = reset_index(Aus_new)

# View first 10 rows of SP500
SP500.head()

# View first 10 rows of NASDAQ
Nasdaq.head()

# View first 10 rows of DJI
DJI.head()

# View first 10 rows of DAX
DAX.head()

# View first 10 rows of CAC40
Paris.head()

# View first 10 rows of Nikkei225
Tokyo.head()

# View first 10 rows of HSI
HongKong.head()

# View first 10 rows of ASX200
Aus.head()

# Export our newly created dataframes as CSV files to the working directory
SP500.to_csv('SP500_2.csv', index=False)
Nasdaq.to_csv('Nasdaq_2.csv', index=False)
DJI.to_csv('DJI_2.csv', index=False)
DAX.to_csv('DAX_2.csv', index=False)
Paris.to_csv('Paris_2.csv', index=False)
Tokyo.to_csv('Tokyo_2.csv', index=False)
HongKong.to_csv('HongKong_2.csv', index=False)
Aus.to_csv('Aus_2.csv', index=False)
